+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **\[component\]** Add the method `AbstractContainer.shouldDisposeEntryOnDispose()` that controls
   whether an entity is detached or disposed when the container is disposed. 
+* **\[component\]** Add invariant check to `AbstractContainer.attach(entity)` to ensure that entity
+  is not already attached.
 
 ### [v0.89](https://github.com/arez/arez/tree/v0.89) (2018-06-07)
 [Full Changelog](https://github.com/arez/arez/compare/v0.88...v0.89)